MGDOBR-1003: reorganization of InstanceTest ... the file was too long
